+++ NEWS.md
@@ -1,6 +1,6 @@
 # ps (development version)
 
-* New `ps_apss()` function to list all running applications on macOS.
+* New `ps_apps()` function to list all running applications on macOS.
 
 * New function `ps_disk_io_counters()` to query disk I/O counters
   (#145, @michaelwalshe).
@@ -23,6 +23,10 @@
   `TERM` signal, and waits for the processes to quit gracefully, via
   `ps_wait()`. The processes that are still alive after the grace period
   are then killed with `SIGKILL`.
+
+* `ps_status()` (and thus `ps()`) is now better at getting the correct
+  status of processes. This usually requires calling the external `ps`
+  tool. See `?ps_status()` on how to opt out from the new behavior (#31).
 
 # ps 1.7.7